requirements_sqlite.txt
Flask==3.1.1
gunicorn==21.2.0
SQLAlchemy==2.0.41
python-dotenv==1.1.1
Werkzeug==3.1.3
Pillow==11.3.0
Flask-Cors==6.0.1
requests
